espresso tests refactoring allure report fix
