+++ jdi-light-mobile-tests/src/test/java/nativeapp_android_espresso/ApiDemosEspressoTestInit.java
@@ -1,6 +1,7 @@
-package nativeapp_android;
+package nativeapp_android_espresso;
 
 import com.epam.jdi.light.driver.WebDriverFactory;
+import com.epam.jdi.light.mobile.AllurePropertiesGenerator;
 import com.epam.jdi.light.mobile.elements.common.AppManager;
 import nativeapp.android.apidemosespresso.ApiDemosEspressoApp;
 import org.testng.annotations.AfterClass;
@@ -25,6 +26,7 @@ public class ApiDemosEspressoTestInit {
 
     @AfterClass(alwaysRun = true)
     public void tearDown() {
+        new AllurePropertiesGenerator().createAllureProperties();
         WebDriverFactory.quitDriverNativeApp();
     }
 }
