+++ jdi-light-mobile-tests/src/test/resources/androidAppsEspresso.xml
@@ -3,7 +3,7 @@
 <suite name="JDI Light Mobile Android Native Apps Espresso Tests" thread-count="1">
     <test name="Android native app Espresso tests">
         <classes>
-            <class name="nativeapp_android.tests.EspressoTests"></class>
+            <class name="nativeapp_android_espresso.tests.EspressoTests"></class>
         </classes>
     </test>
 </suite>
